Tried to finish my part
Hey. When I read about the pure sine wave correlation, I do not know if it is completely ok, but I think they are equeal. The part with phases got cancelled in the formula of covariance. Anyways, I will put this here so please merge or copy the text I put there. I made some changes to the synthesis block, so now looks a bit different, but enough ok I think, or I do not know. It looks so different, matlab does not let me hear the sounds clearly, the media players I have do not let me neither, this is sad. It at least sounds like an old noisy radio, so... At least it works. Bad, but works. Anyways, Happy holidays, and please send the file :)